Fix quickstart Anthropic key guard
